pyproject.toml
[tool.poetry]
name = "da-ayr-beta-webapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.bandit]
exclude_dirs = ['*_test.py', '*/test_*.py', 'venv', 'node_modules']
skips = ["B105"]

[tool.poetry.dependencies]
python = "^3.11"
cssmin = "0.2.0"
email-validator = "2.1.0.post1"
flask-assets = "2.1.0"
flask-compress = "1.14"
flask-limiter = "3.5.0"
flask-talisman = "1.1.0"
flask = "2.3.3"
govuk-frontend-jinja = "2.7.0"
govuk-frontend-wtf = "2.5.0"
gunicorn = "21.2.0"
jsmin = "3.0.1"
python-dotenv = "1.0.0"
redis = "5.0.1"
python-keycloak = "3.7.0"
pytest-playwright = "0.4.3"
flask-pytest = "^0.0.5"
boto3 = "1.28.64"
botocore = "1.31.85"
opensearch-py = "2.3.2"
bs4 = "^0.0.1"
pytest-cov = "^4.1.0"
moto = "^4.2.8"
sqlalchemy = "^2.0.23"
flask-sqlalchemy = "^3.1.1"
pytest-flask-sqlalchemy = "^1.1.0"
testing-postgresql = "^1.3.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
bandit = "1.7.5"
black = "23.11.0"
flake8-bugbear = "23.9.16"
isort = "5.12.0"
pep8-naming = "0.13.3"
pip-tools = "7.3.0"
pur = "7.3.1"
pytest-cov = "4.1.0"
safety = "2.3.4"
pytest-flask = "^1.2.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.isort]
line_length = 80
